pyproject.toml
[tool.poetry]
name = "jalluindeksoija"
version = "0.1.0"
description = ""
authors = ["Eetu Raunio <eetu.raunio@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.11"
gql = {extras = ["aiohttp"], version = "^3.1.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
